moved RoundData to src, dividing the latestIndex, updated validation and tests
