--- typings/reference-data-reader/index.d.ts
@@ -1,9 +1,0 @@
-declare module 'reference-data-reader' {
-  export interface RoundData {
-    roundId: BigNumber
-    answer: BigNumber
-    startedAt: BigNumber
-    updatedAt: BigNumber
-    answeredInRound: BigNumber
-  }
-}
